TST: series/indexing tests parametrization + moving test methods (#20059)
